refactored the BenefitExpansion component
